pybk/requirements.txt
celery[redis,auth,msgpack]>=4.4.7
flask>=1.1.2
flask-cors>=3.0.10
flask-swagger-ui==3.36.0
gunicorn>=20.1.0
gevent
pyyaml>=5.4.1
supervisor>=4.2.2
numpy>=1.26.3
librosa>=0.8.0
webrtcvad>=2.0.10
matplotlib
memory-tempfile
scipy
scikit-learn
spafe
pydub
python_speech_features
redis
speechbrain==1.0.0
torchaudio==2.2.1
